- Change model, animation, flex animation and vertex animation UDM storage formats - Add new UDM Lua bindings - Add animation save optimization to remove redundant animation data - Change bone id type from uint32 to uint16 (to save space)

+++ core/shared/include/pragma/entities/components/base_animated_component.hpp
@@ -20,7 +20,7 @@ class Animation;
 class Frame;
 class ModelSubMesh;
 struct AnimationEvent;
-using BoneId = uint32_t;
+using BoneId = uint16_t;
 enum class ALSoundType : int32_t;
 namespace pragma
 {
